requirements.txt
openai-whisper
python-docx
ffmpeg-python
reportlab
